+++ src/main/java/ch/uzh/ifi/seal/soprafs20/constant/Category.java
@@ -1,0 +1,5 @@
+package ch.uzh.ifi.seal.soprafs20.constant;
+
+public enum Category {
+    // TODO
+}
