+++ src/main/java/ch/uzh/ifi/seal/soprafs20/constant/Mode.java
@@ -1,0 +1,5 @@
+package ch.uzh.ifi.seal.soprafs20.constant;
+
+public enum Mode {
+    SOCIAL, QUICK, SINGLE;
+}
