+++ src/main/java/ch/uzh/ifi/seal/soprafs20/controller/UserController.java
@@ -18,67 +18,31 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-enum mode{
-    Social(0), Quick(1), Single(2);
-
-    private int mode;
-
-    mode(int Stat){
-        this.mode = Stat;
-    }
-
-    public int getMode(){
-        return mode;
-    }
-}
-
-enum category{
-    HP(0), ATK(1), Wheight(2);
-
-    private int mode;
-
-    category(int Stat){
-        this.mode = Stat;
-    }
-
-    public int getCategory(){
-        return mode;
-    }
-}
-
-
 @RestController
 public class UserController {
 
     private final UserService userService;
 
-    private final GameService gameService;
-
-    private final CardService cardService;
-
-
     UserController(UserService userService, GameService gameService, CardService cardService) {
         this.userService = userService;
-        this.gameService = gameService;
-        this.cardService = cardService;
     }
 
 
-    /**     #1      **/
+    /*     #1      **/
     /** This request creates a new user, if the username is not already used **/
     @PostMapping("/users")
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
-    public void createUser(@RequestBody UserPostDTO userPostDTO, @RequestHeader("Token") String token) {
+    public void createUser(@RequestBody UserPostDTO userPostDTO) {
         // convert API user to internal representation
         User userInput = DTOMapper.INSTANCE.convertUserPostDTOtoEntity(userPostDTO);
 
         // create user
-        User createdUser = userService.createUser(userInput);
+        userService.createUser(userInput);
     }
 
 
-    /**     #2      **/
+    /*     #2      **/
     /** This request returns a list with all users **/
     @GetMapping("/users")
     @ResponseStatus(HttpStatus.CREATED)
@@ -97,7 +61,7 @@ public class UserController {
 
 
 
-    /**     #3      **/
+    /*     #3      **/
     /** This request returns a User object corresponding to the input id **/
     @GetMapping("/users/{userId}")
     @ResponseStatus(HttpStatus.OK)
@@ -111,7 +75,7 @@ public class UserController {
     }
 
 
-    /**     #4      **/
+    /*     #4      **/
     /** This request updates a User **/
     @PutMapping("/users/{userId}")
     @ResponseStatus(HttpStatus.OK)
@@ -124,12 +88,12 @@ public class UserController {
     }
 
 
-    /**     #5      **/
+    /*     #5      **/
     /** This request logs in a user if the credentials are correct **/
     @PutMapping("/login")
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public String userLogin(@RequestBody UserPutDTO userPutDTO, @RequestHeader("Token") String token) {
+    public String userLogin(@RequestBody UserPutDTO userPutDTO) {
         // convert API user to internal representation
         User userInput = DTOMapper.INSTANCE.convertUserPutDTOtoEntity(userPutDTO);
         String tokenNr = userInput.getToken();
@@ -145,7 +109,7 @@ public class UserController {
     }
 
 
-    /**     #6      **/
+    /*     #6      **/
     /** This request logs  out a user **/
     @PutMapping("/logout")
     @ResponseStatus(HttpStatus.OK)
@@ -154,126 +118,4 @@ public class UserController {
         userService.logUserOut(token);
     }
 
-
-    /**     #7      **/
-    /** This request creates a new lobby and returns the gameToken for sharing**/
-    @PostMapping("/games")
-    @ResponseStatus(HttpStatus.CREATED)
-    @ResponseBody
-    public String createLobby(String gameName, Enum mode, @RequestHeader("Token") String token) {
-
-        //creates a game with the given gamemode and gamename
-        String newGameToken = gameService.createGame(gameName,mode);
-
-        return newGameToken;
-    }
-
-
-    /**     #8      **/
-    /** This request let's a user join or leave a lobby or kick another player from the lobby **/
-    @PutMapping("/games/{gameToken}/users")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public void lobbyOperation(@PathVariable String gameToken, String userName, Integer action, @RequestHeader("Token") String token ) {
-
-        gameService.addUser(userName,gameToken);
-
-        /**
-        long value = action;
-
-        if(value==0){
-            gameService.addUser(userName,gameToken);
-        }
-        else if(value==1){
-            // for leaving, the own username was inputted
-            gameService.removeUser(userName,gameToken);
-        }
-        else if(value==2){
-            gameService.addUser(userName,gameToken);
-        }**/
-    }
-
-
-    /**     #9      **/
-    /** This request starts a game from an existing lobby according to the gameToken **/
-    @PutMapping("/games/{gameToken}")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public void startGame(@PathVariable String gameToken, Integer amount, @RequestHeader("Token") String token) {
-        //sets amount of npcs and starts the game
-        gameService.setNPCAmount(amount, gameToken);
-    }
-
-
-
-    /**     #10     **/
-    /** This request returns the current state of a running game **/
-    @GetMapping("/games/{gameToken}")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public TableGetDTO getGameState(@PathVariable String gameToken, @RequestHeader("Token") String token) {
-
-        //gets gamestate as a table object
-        Tables table = gameService.getGame(gameToken);
-
-        //convert to correct API format to return
-        return DTOMapper.INSTANCE.convertEntityToTableGetDTO(table);
-    }
-
-
-    /**     #11     **/
-    /** This request lets the user, whose turn it is, select a category for battle **/
-    @PutMapping("/games/{gameToken}/categories")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public void selectAttribute(@PathVariable String gameToken, Enum category, @RequestHeader("Token") String token) {
-
-        //call gameservice method for selecting a category
-        gameService.selectCategory(category, gameToken);
-
-    }
-
-
-    /**     #12     **/
-    /** This request lets a card evolve **/
-    @PutMapping("/games/{gameToken}/berries")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public void berryUpgrade(@PathVariable String gameToken, Integer amount, String userName, @RequestHeader("Token") String token) {
-
-        //call gameservice method for berry usage
-        gameService.userBerries(amount, userName);
-    }
-
-    /**     #13     **/
-    /** This request returns a map with all Pokémon IDs and Sprite urls **/
-
-    @GetMapping("/cards")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public Map<Long,String> getCardList(@RequestHeader("Token") String token) {
-
-        //get a map with all ids and urls of the cards
-        Map<Long,String> map = cardService.getCards();
-
-        return map;
-    }
-
-
-    /**     #14     **/
-    /** This request returns a complete pkm-card **/
-
-    @GetMapping("/cards/{pokemonId}")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public CardGetDTO getCard(@PathVariable String pokemonId, @RequestHeader("Token") String token) {
-
-        //convert pathvaraible to long
-        long cardId = Long.parseLong(pokemonId);
-
-        //get card from service
-        Card card = cardService.getCard(cardId);
-
-        return DTOMapper.INSTANCE.convertEntityToCardGetDTO(card);
-    }
 }
